`Composite Key` test & implementation
A composite key is a key that is composed of more than one attribute.
In this case, we have a Category class that has a name and a section attribute.
Which can help us to identify a category.

For example, "Java" could be both in the "Programming" and "Geography" section (Java could be an island of Indonesia).

So we need to use composite keys to differentiate each of them.